Make timeout dependent on important engines
Important engines are engines w/ a WEIGHT of above 1. These engines are
always allows to finish regardless of timeouts. All other engines that
finish while the important engines are running are also not under any
timeout. Only when the important engines are finished, then the
remaining engines are restricted to running an additional 0.5 seconds.